Floating enemy pos + speed added

+++ Deep Down/Motor2D/CatPeasant.cpp
@@ -12,7 +12,7 @@
 
 #include "SDL/include/SDL_timer.h"
 
-CatPeasant::CatPeasant(int x, int y, PathInfo* path) : Enemy(x, y, path)
+CatPeasant::CatPeasant(float x, float y, PathInfo* path) : Enemy(x, y, path)
 {
 	idle.PushBack({ 2, 21, 64, 64 });
 	idle.PushBack({ 68, 21, 64, 64 });
