+++ Deep Down/Motor2D/Plant.h
@@ -9,7 +9,7 @@ private:
 	Animation idle;
 
 public:
-	Plant(int x, int y, PathInfo* path);
+	Plant(float x, float y, PathInfo* path);
 	void OnCollision(Collider* c1, Collider* c2);
 	void Move();
 };
